+++ BalanceApi/Controllers/ViewModels/AddAccountTypeRequest.cs
@@ -1,12 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BalanceApi.Controllers.Views.Request
+namespace BalanceApi.Controllers.ViewModels
 {
-    public class UpdateAccountType
+    public class AddAccountType
     {
-        [Required]
-        public long Id { get; set; }
-        
         [Required]
         [StringLength(100)]
         public string Name { get; set; }
