+++ BalanceApi/Controllers/ViewModels/UpdateAccountTypeRequest.cs
@@ -1,9 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BalanceApi.Controllers.Views.Request
+namespace BalanceApi.Controllers.ViewModels
 {
-    public class AddAccountType
+    public class UpdateAccountType
     {
+        [Required]
+        public long Id { get; set; }
+        
         [Required]
         [StringLength(100)]
         public string Name { get; set; }
